Beta 2

+++ Program.cs
@@ -4,35 +4,37 @@ namespace TgSupportBot;
 
 public static class Program
 {
-    private const string LogPath = "./latest.log";
-    private static readonly FileInfo TokenFile = new("./token.txt");
+	private const string LogPath = "./latest.log";
+	private static readonly FileInfo TokenFile = new("./token.txt");
 
-    public static async Task<int> Main()
-    {
-        if (File.Exists(LogPath))
-        {
-            File.Delete(LogPath);
-        }
-        Log.Logger = new LoggerConfiguration()
-            #if DEBUG
-            .MinimumLevel.Verbose()
-            #endif
-            .WriteTo.Console()
-            .WriteTo.File(LogPath)
-            .CreateLogger();
+	public static async Task<int> Main(string[] args)
+	{
+		if (File.Exists(LogPath) && !args.Contains("--log-append"))
+		{
+			File.Delete(LogPath);
+		}
+		Log.Logger = new LoggerConfiguration()
+#if DEBUG
+			.MinimumLevel.Verbose()
+#else
+			.MinimumLevel.Information()
+#endif
+			.WriteTo.Console()
+			.WriteTo.File(LogPath)
+			.CreateLogger();
 
-        try
-        {
-            string token = await File.ReadAllTextAsync(TokenFile.FullName);
-            
-            BotEngine metBot = new(token);
-            await metBot.ListenForMessagesAsync();
-        }
-        catch (Exception e)
-        {
-            Log.Logger.Fatal(e, "An error occured during bot initialization");
-        }
+		try
+		{
+			string token = await File.ReadAllTextAsync(TokenFile.FullName);
+			
+			BotEngine metBot = new(token);
+			await metBot.Start();
+		}
+		catch (Exception e)
+		{
+			Log.Logger.Fatal(e, "Uncaughted exception during bot work.");
+		}
 
-        return 0;
-    }
+		return 0;
+	}
 }
